Fix clipping performance regression.
The oceans test is an outlier because it has unusually large polygons
that are tested against every single ring to see if the first point of a
ring is contained (and thus the whole ring), to ensure correct polygon
semantics.

However, in spherical geometry, it is equivalent to test whether the
exterior is inside the ring, assuming the ring has the correct winding
order, hence we can improve performance by picking the smallest ring as
the polygon for the point-in-polygon routine.